Adds doorkeeper oauth generation
This commit adds gems cancancan (which will be used more in the future for authorization on other things like mod/admin roles, race creator permissions) and doorkeeper, which automates oauth code generation.  The websocket connection will accept either no code and pull from cookies with warden, no authentication for anon connection, or a valid oauth token.  If an invalid token is supplied to the websocket connection it will reject it.  This also opens up the websocket to all hosts instead of just the base one.  There is no documentation on oauth generation, that will come at a later time.